Default configurators do more for distros
