+++ opentelemetry-distro/setup.cfg
@@ -52,7 +52,7 @@ where = src
 opentelemetry_distro =
     distro = opentelemetry.distro:OpenTelemetryDistro
 opentelemetry_configurator =
-    configurator = opentelemetry.distro:Configurator
+    configurator = opentelemetry.distro:OpenTelemetryConfigurator
 
 [options.extras_require]
 test =
