+++ opentelemetry-distro/src/opentelemetry/distro/__init__.py
@@ -14,143 +14,16 @@
 #
 import os
 from logging import getLogger
-from os import environ
-from typing import Sequence, Tuple
 
-from pkg_resources import iter_entry_points
-
-from opentelemetry import trace
-from opentelemetry.environment_variables import (
-    OTEL_PYTHON_ID_GENERATOR,
-    OTEL_TRACES_EXPORTER,
-)
-from opentelemetry.instrumentation.configurator import BaseConfigurator
+from opentelemetry.environment_variables import OTEL_TRACES_EXPORTER
 from opentelemetry.instrumentation.distro import BaseDistro
-from opentelemetry.sdk.resources import SERVICE_NAME, Resource
-from opentelemetry.sdk.trace import TracerProvider
-from opentelemetry.sdk.trace.export import BatchSpanProcessor, SpanExporter
-from opentelemetry.sdk.trace.id_generator import IdGenerator
+from opentelemetry.sdk._configuration import _OTelSDKConfigurator
 
 logger = getLogger(__file__)
 
 
-EXPORTER_OTLP = "otlp"
-EXPORTER_OTLP_SPAN = "otlp_proto_grpc_span"
-
-RANDOM_ID_GENERATOR = "random"
-_DEFAULT_ID_GENERATOR = RANDOM_ID_GENERATOR
-
-
-def _get_id_generator() -> str:
-    return environ.get(OTEL_PYTHON_ID_GENERATOR, _DEFAULT_ID_GENERATOR)
-
-
-def _get_exporter_names() -> Sequence[str]:
-    trace_exporters = environ.get(OTEL_TRACES_EXPORTER)
-
-    exporters = set()
-
-    if trace_exporters and trace_exporters.lower().strip() != "none":
-        exporters.update(
-            {
-                trace_exporter.strip()
-                for trace_exporter in trace_exporters.split(",")
-            }
-        )
-
-    if EXPORTER_OTLP in exporters:
-        exporters.remove(EXPORTER_OTLP)
-        exporters.add(EXPORTER_OTLP_SPAN)
-
-    return list(exporters)
-
-
-def _init_tracing(
-    exporters: Sequence[SpanExporter], id_generator: IdGenerator
-):
-    # if env var OTEL_RESOURCE_ATTRIBUTES is given, it will read the service_name
-    # from the env variable else defaults to "unknown_service"
-    provider = TracerProvider(
-        id_generator=id_generator(),
-    )
-    trace.set_tracer_provider(provider)
-
-    for _, exporter_class in exporters.items():
-        exporter_args = {}
-        provider.add_span_processor(
-            BatchSpanProcessor(exporter_class(**exporter_args))
-        )
-
-
-def _import_tracer_provider_config_components(
-    selected_components, entry_point_name
-) -> Sequence[Tuple[str, object]]:
-    component_entry_points = {
-        ep.name: ep for ep in iter_entry_points(entry_point_name)
-    }
-    component_impls = []
-    for selected_component in selected_components:
-        entry_point = component_entry_points.get(selected_component, None)
-        if not entry_point:
-            raise RuntimeError(
-                "Requested component '{}' not found in entry points for '{}'".format(
-                    selected_component, entry_point_name
-                )
-            )
-
-        component_impl = entry_point.load()
-        component_impls.append((selected_component, component_impl))
-
-    return component_impls
-
-
-def _import_exporters(
-    exporter_names: Sequence[str],
-) -> Sequence[SpanExporter]:
-    trace_exporters = {}
-
-    for (
-        exporter_name,
-        exporter_impl,
-    ) in _import_tracer_provider_config_components(
-        exporter_names, "opentelemetry_exporter"
-    ):
-        if issubclass(exporter_impl, SpanExporter):
-            trace_exporters[exporter_name] = exporter_impl
-        else:
-            raise RuntimeError(
-                "{0} is not a trace exporter".format(exporter_name)
-            )
-    return trace_exporters
-
-
-def _import_id_generator(id_generator_name: str) -> IdGenerator:
-    # pylint: disable=unbalanced-tuple-unpacking
-    [
-        (id_generator_name, id_generator_impl)
-    ] = _import_tracer_provider_config_components(
-        [id_generator_name.strip()], "opentelemetry_id_generator"
-    )
-
-    if issubclass(id_generator_impl, IdGenerator):
-        return id_generator_impl
-
-    raise RuntimeError("{0} is not an IdGenerator".format(id_generator_name))
-
-
-def _initialize_components():
-    exporter_names = _get_exporter_names()
-    trace_exporters = _import_exporters(exporter_names)
-    id_generator_name = _get_id_generator()
-    id_generator = _import_id_generator(id_generator_name)
-    _init_tracing(trace_exporters, id_generator)
-
-
-class Configurator(BaseConfigurator):
-
-    # pylint: disable=no-self-use
-    def _configure(self, **kwargs):
-        _initialize_components()
+class OpenTelemetryConfigurator(_OTelSDKConfigurator):
+    pass
 
 
 class OpenTelemetryDistro(BaseDistro):
